+++ contracts/updatecustodianmultisigwallet/UpdateCustodianMultisigWallet.sol
@@ -1,5 +1,4 @@
 pragma gosh-solidity >=0.76.1;
-pragma ignoreIntOverflow;
 pragma AbiHeader expire;
 pragma AbiHeader pubkey;