+++ http-server/src/metrics.rs
@@ -9,7 +9,7 @@ pub struct RoutingMetrics {
     ext_msg_delivery_duration: Histogram<u64>,
     ext_msg_processing_duration: Histogram<u64>,
     boc_by_address_response: Histogram<u64>,
-    block_request: Counter<u64>,
+    state_request: Counter<u64>,
 }
 
 impl RoutingMetrics {
@@ -30,7 +30,7 @@ impl RoutingMetrics {
                 .u64_histogram("node_boc_by_address_response")
                 .with_boundaries(boundaries)
                 .build(),
-            block_request: meter.u64_counter("node_block_req_recv").build(),
+            state_request: meter.u64_counter("node_state_req_recv").build(),
         }
     }
 
@@ -50,7 +50,7 @@ impl RoutingMetrics {
         self.boc_by_address_response.record(value, &[KeyValue::new("code", http_code.to_string())]);
     }
 
-    pub fn report_block_request(&self) {
-        self.block_request.add(1, &[]);
+    pub fn report_state_request(&self) {
+        self.state_request.add(1, &[]);
     }
 }