+++ node/src/block/producer/builder/build_actions.rs
@@ -73,8 +73,6 @@ use super::PreparedBlock;
 use super::ThreadResult;
 use crate::block::postprocessing::postprocess;
 use crate::block::producer::builder::trace::simple_trace_callback;
-use crate::block::producer::errors::verify_error;
-use crate::block::producer::errors::BP_DID_NOT_PROCESS_ALL_MESSAGES_FROM_PREVIOUS_BLOCK;
 use crate::block::producer::execution_time::ExecutionTimeLimits;
 use crate::block::producer::wasm::WasmNodeCache;
 use crate::block_keeper_system::epoch::decode_epoch_data;
@@ -2067,14 +2065,15 @@ impl BlockBuilder {
         active_destinations: &HashMap<AccountAddress, u128>,
         check_messages_map: &mut Option<HashMap<AccountAddress, BTreeMap<u64, UInt256>>>,
         there_are_no_new_messages_for_verify_block: &mut bool,
-        verify_block_contains_missing_messages_from_prev_state: bool,
+        _verify_block_contains_missing_messages_from_prev_state: bool,
     ) -> anyhow::Result<Option<(Message, AccountAddress, u128, OutMsgQueueKey)>> {
         for (index, (message, _tr_cell)) in &self.new_messages {
             let acc_id = message.int_dst_account_id().unwrap_or_default().into();
             if let Some(msg_set) = &check_messages_map {
-                if verify_block_contains_missing_messages_from_prev_state {
-                    return Err(verify_error(BP_DID_NOT_PROCESS_ALL_MESSAGES_FROM_PREVIOUS_BLOCK));
-                }
+                // TODO: This check seems to be wrong now. Check it and unlock
+                // if verify_block_contains_missing_messages_from_prev_state {
+                //     return Err(verify_error(BP_DID_NOT_PROCESS_ALL_MESSAGES_FROM_PREVIOUS_BLOCK));
+                // }
                 let Some(messages) = msg_set.get(&acc_id) else {
                     continue;
                 };